Added ability to check multiple branches
Using the `branches` key under `source` turns on multi-branch mode.
Branches takes a regular expression pattern, and along with `ignore_branches`,
can be used to limit checking to a specific set of branches.

This will detect changes on existing branches as well as new branches.
Only changes on the same branch will be grouped together in the check,
so changes on different branches will be detected on subsequent checks.

Also:

* Added cleanup trap on exit, improved messaging

There is now a trap on script exit that detects if the script exited due
to the effects of `set -e` and prints a failure message.  This trap is
also where the temp directory gets deleted -- failures meant the
directory stayed around due to premature exit.  In order to assist
debugging, the directory is still kept, but a message to delete it when
done is now displayed.

* Fixed 'Too many arguements' error

The tests were showing the following error:
  /opt/resource/check: line 49: [: too many arguments

To resolve this, the tested arguments were wrapped in quotes.  All tests
passed with this change.

* Refactored check_* and get_* test helper functions into single functions

In order to better support the many combinations of arguments to check,
the check_* functions were refactored into a single check() function
that parse pairs of arguments and assemble the correct json using jq
merge operator (*)

This was made necessary due to the multi_branch option, which would have
doubled the existing 9 check_* functions.

+++ test/all.sh
@@ -4,10 +4,25 @@ set -e
 
 export TMPDIR_ROOT=$(mktemp -d /tmp/git-tests.XXXXXX)
 
+on_exit() {
+  exitcode=$?
+  if [ $exitcode != 0 ] ; then
+    echo -e '\e[41;33;1m'"Failure encountered!"'\e[0m'
+    echo ""
+    echo "Delete $TMPDIR_ROOT when done inspecting the failure"
+    echo ""
+  else
+    rm -rf $TMPDIR_ROOT
+  fi
+
+}
+
+trap on_exit EXIT
+
 $(dirname $0)/check.sh
+$(dirname $0)/check_multi.sh
 $(dirname $0)/get.sh
 $(dirname $0)/put.sh
 
-echo -e '\e[32mall tests passed!\e[0m'
+echo -e '\e[32;1m'"All tests passed!"'\e[0m'
 
-rm -rf $TMPDIR_ROOT
